+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] -  2021-01-18
+
+## Added
+
+- [#6](https://github.com/creekorful/srcode/issues/6) Implement srcode set-cmd.
+- [#7](https://github.com/creekorful/srcode/issues/7) Implement srcode mv.
+
+## Changed
+
+- Improve documentation.
+- Improve output of srcode ls.
+
+## [0.2.0] - 2021-01-17
+
 ## Added
 
 - Now create README.md in meta directory to explain how to use it.
@@ -26,5 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial pre-release.
 
-[Unreleased]: https://github.com/creekorful/srcode/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/creekorful/srcode/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/creekorful/srcode/compare/v0.3.0...HEAD
+[0.2.0]: https://github.com/creekorful/srcode/compare/v0.2.0...HEAD
 [0.1.0]: https://github.com/creekorful/srcode/releases/tag/v0.1.0